# subscription-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "https://www.cloudbees.com/products/terms-service",
        "contact": {
            "name": "CloudBees Support",
            "url": "http://support.cloudbees.com",
            "email": "support@cloudbees.com"
        },
        "license": {}
    },
    "host": "localhost:8085",
    "basePath": "/api/v1",
    "paths": {
        "/accounts": {
            "get": {
                "description": "Gets an array of accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetAccounts",
                "operationId": "cloud-bill-saas-subscription-service-get-accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "optional comma separated list of filter",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "optional order",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/persistence.Account"
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Upsert an account passing account json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upsert an account",
                "operationId": "cloud-bill-saas-subscription-service-upsert-account",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/persistence.Account"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Upserted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "persistence.Account": {
            "type": "object",
            "properties": {
                "approvals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/persistence.Approval"
                    }
                },
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
        "persistence.Approval": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
    }
}